requirements.txt
google-adk
google-cloud-aiplatform
google-genai
python-docx
pypdf
streamlit